Integrate action files and delete unnecessary email
