+++ src/app/actions/reset_password.tsx
@@ -1,4 +1,4 @@
-import { EmailforPassReset, ResponseData } from "../types/types";
+import { EmailforPassReset, NewPasswordData, ResponseData } from "../types/types";
 
 const API_BASE_URL = process.env.NEXT_PUBLIC_API_BASE_URL;
 
@@ -25,6 +25,33 @@ export const verifyEmail = async (
 
 	} catch (error) {
 		console.error("Error during password reset request:", error);
+		return { success: false, data: "An error occurred" };
+	}
+};
+
+export const setNewPassword = async (
+	newPasswordData: NewPasswordData
+): Promise<ResponseData<string>> => {
+	try {
+		const response = await fetch(`${API_BASE_URL}/api/new_password`, { // TBD
+			method: "POST",
+			headers: {
+				"Content-Type": "application/json",
+			},
+			body: JSON.stringify({
+				passwordHash: newPasswordData.passwordFirst,
+			}),
+		})
+			.then((res) => {
+				return res.json();
+			})
+			.then((data: ResponseData<string>) => {
+				return data;
+			});
+
+		return response;
+	} catch (error) {
+		console.error("Error during setting new passowrd:", error);
 		return { success: false, data: "An error occurred" };
 	}
 };
